@@ -68,41 +55,45 @@
     "@swc/core": "1.2.155",
     "@swc/helpers": "~0.3.3",
     "@swc/jest": "0.2.20",
-    "@testing-library/react": "12.1.3",
+    "@testing-library/react": "13.0.0",
     "@testing-library/react-hooks": "7.0.2",
     "@types/jest": "27.4.1",
-    "@types/node": "^17.0.21",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
+    "@types/node": "^17.0.24",
+    "@types/react": "18.0.0",
+    "@types/react-dom": "18.0.0",
     "@types/react-is": "17.0.3",
     "@types/react-router-dom": "5.3.3",
-    "@types/styled-components": "5.1.24",
+    "@types/styled-components": "5.1.25",
     "@types/uuid": "^8.3.4",
-    "@typescript-eslint/eslint-plugin": "~5.13.0",
-    "@typescript-eslint/parser": "~5.10.0",
+    "@typescript-eslint/eslint-plugin": "5.18.0",
+    "@typescript-eslint/parser": "5.18.0",
     "@vitejs/plugin-react-refresh": "^1.3.6",
+    "babel-eslint": "^10.0.0",
     "babel-jest": "27.5.1",
     "babel-loader": "8.2.3",
+    "babel-plugin-module-resolver": "4.1.0",
     "babel-plugin-styled-components": "2.0.6",
+    "babel-preset-gatsby": "2.2.0",
+    "commitizen": "^4.2.4",
     "csstype": "^3.0.10",
     "cypress": "^9.1.0",
-    "discord-api-types": "^0.31.0",
+    "cz-conventional-changelog": "^3.3.0",
     "dotenv": "^16.0.0",
     "esbuild": "^0.14.26",
-    "eslint": "~8.11.0",
-    "eslint-config-next": "12.1.1-canary.3",
+    "eslint": "8.12.0",
     "eslint-config-prettier": "8.4.0",
     "eslint-plugin-cypress": "^2.10.3",
-    "eslint-plugin-import": "2.25.4",
+    "eslint-plugin-import": "2.26.0",
     "eslint-plugin-jsx-a11y": "6.5.1",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "4.0.0",
     "eslint-plugin-react": "7.29.4",
-    "eslint-plugin-react-hooks": "4.3.0",
+    "eslint-plugin-react-hooks": "4.4.0",
     "husky": "^7.0.4",
     "jest": "^27.5.1",
     "jest-svg-transformer": "^1.0.0",
     "ngx-deploy-npm": "^3.1.2",
+    "nx": "13.10.2",
     "nx-plugin-vite": "1.5.1",
     "prettier": "^2.5.1",
     "react-is": "18.0.0",
