@@ -9,19 +9,21 @@ export const CoreButton = styled(motion.button)`
 
   background-color: transparent;
 
-  padding: 0.5rem 1rem;
+  padding: 0.8rem 1.1rem;
   border-radius: 4px;
   /* transition: background-color 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms,
     box-shadow 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms, border-color 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms,
     color 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms; */
   transition: all 0.3s ease;
-  font-size: ${({ size }: Props) => `${size || '1.05'}rem`};
+  font-size: ${({ size }: Props) => {
+    return `${size || '1.05'}rem`;
+  }} !important;
   font-weight: 500;
   cursor: pointer;
   position: relative;
   overflow: hidden;
 
-  line-height: 1.75;
+  /* line-height: 1.75; */
 
   @media screen and (max-width: 600px) {
     font-size: ${({ size }: Props) => {
