@@ -1,7 +1,10 @@
-import { Title } from '@v-thomas/shared/ui';
+import { Button, Title } from '@v-thomas/shared/ui';
 import { NavLinks, Navbar as SNavbar } from './navbar.styles';
 import { SignInButton } from '../../auth/sign-in-button/sign-in-button';
 import { SignUpButton } from '../../auth/sign-up-button/sign-up-button';
+import { useUser } from '@v-thomas/root/libs/thunder/hooks/src/lib';
+import { SignOutButton } from '../../auth';
+import { useNavigate } from 'react-router-dom';
 
 const NavbarAni = {
   initial: { opacity: 0, y: -10 },
@@ -15,14 +18,27 @@ const NavbarAni = {
 };
 
 export function PublicNavbar() {
+  const { isAuthed } = useUser();
+
+  const router = useNavigate();
+
   return (
     <SNavbar>
       <Title size="1.8" variants={NavbarAni} initial="initial" animate="animate" exit="exit">
         Movie Reviewer
       </Title>
       <NavLinks variants={NavbarAni} initial="initial" animate="animate" exit="exit">
-        <SignInButton />
-        <SignUpButton />
+        {isAuthed ? (
+          <>
+            <SignOutButton />
+            <Button onClick={() => router('/app')}>To Lobby</Button>
+          </>
+        ) : (
+          <>
+            <SignInButton />
+            <SignUpButton />
+          </>
+        )}
       </NavLinks>
     </SNavbar>
   );
